requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.6
django-allauth==0.50.0
django-cors-headers==3.13.0
django-rest-swagger==2.0.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
exponent-server-sdk==2.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.9.0
sqlparse==0.4.2
twilio==7.12.0
uritemplate==4.1.1
urllib3==1.26.10
youtube-dl==2021.12.17
